requirements.txt
streamlit
ultralytics
torch
torchvision
opencv-python-headless
numpy
Pillow
pandas
pyyaml
